sim: frv: flip trapdump default back to off
When I refactored this by scoping it to sim-frv-xxx in commit
aba1d3c5ecbaf371cd399b9a7cf59f9aa7c906d6 ("sim: frv: scope the
unique configure flag"), I changed the default from off to on.
While the feature is nice for developers, it breaks a bunch of
tests which aren't expecting this extra output.  So flip it back
to off by default.